Add files via upload
Changed the SaveReportData method to saveRequest, so it's usable by all methods in the method caller class now. Added a method to get all of the evar,props, and events for a reportsuite. Created a test for each of those new methods. Created a file that will run all of the tests at once inside the TestFilesJSON folder.